refactor: remove helper module
helper module only had a single function that was used by Model. This commit deletes the module and moves it to Model class as a staticmethod.
